fix: extract shared createApp/listenOnRandomPort helpers from buildInfoServerCleanup.test.js (#427)
Closes #427
